Combat system & Inventory - selecting weapon
Player can double click on weapon in inventory menu to put it on.  This make weapon visible on character.  When client select weapon its name is send to server, which validate if this character has this item and set it. Weapon Component is replicated, thanks to every clients see new weapon mesh.