Weapon - bug fix
Weapon as component didn't work correct. There were lot of issues eg properties was disappearing in editor, wasn't attached to socket. There was also problem with setting mesh for weapon, it was because of soft references in data table for weapon. Now instead of component for weapon is actor, which is used in character as child actor and in data table are raw poiters to assets.