Spells - use
Changing selected spell by mouse whell, change now real selected spell, not only UI. Player can use spell. It is prevented for cheating - check if spells set as selected are in chracter inventory, but it not good. It should be changed in the future to check if character has specified spell only when he is using it.